+++ src/surrealist/__init__.py
@@ -1,5 +1,5 @@
 from .connections import WebSocketConnection, HttpConnection, Connection
-from .enums import Algorithm
+from .enums import Algorithm, AutoOrNone
 from .errors import *
 from .ql import Database, DatabaseConnectionsPool, Table, Where
 from .record_id import RecordId
@@ -10,5 +10,5 @@ from .utils import get_uuid, to_surreal_datetime_str, to_datetime, LOG_FORMAT
 __all__ = ("Surreal", "SurrealResult", "WebSocketConnection", "HttpConnection", "PySurrealError", "HttpConnectionError",
            "HttpClientError", "SurrealConnectionError", "WebSocketConnectionError", "WebSocketConnectionClosedError",
            "ConnectionParametersError", "CompatibilityError", "OperationOnClosedConnectionError", "WrongCallError",
-           "Connection", "get_uuid", "Database", "Table", "Where", "DatabaseConnectionsPool",
+           "Connection", "get_uuid", "Database", "Table", "Where", "DatabaseConnectionsPool", "AutoOrNone",
            "to_surreal_datetime_str", "to_datetime", "LOG_FORMAT", "Algorithm", "RecordId", "SurrealRecordIdError")