+++ tests/integration_tests/test_http_connection.py
@@ -286,8 +286,7 @@ class TestHttpConnection(TestCase):
         with surreal.connect() as connection:
             connection.use("test", "test")
             res = connection.count("not_exists")
-            self.assertFalse(res.is_error())
-            self.assertEqual(0, res.result)
+            self.assertTrue(res.is_error(), res)
             self.assertEqual("SELECT count() FROM not_exists GROUP ALL;", res.query)
             count = connection.count("article").result
             uid = get_random_series(6)
@@ -295,13 +294,12 @@ class TestHttpConnection(TestCase):
             new_count = connection.count("article").result
             self.assertEqual(new_count, count + 1)
 
-    def test_count_is_zero_if_wrong(self):
+    def test_count_is_error_if_wrong(self):
         surreal = Surreal(URL, credentials=('root', 'root'), use_http=True)
         with surreal.connect() as connection:
             connection.use("test", "test")
             res = connection.count("wrong")
-            self.assertFalse(res.is_error())
-            self.assertEqual(0, res.result)
+            self.assertTrue(res.is_error())
 
     def test_count_returns_fields(self):
         surreal = Surreal(URL, credentials=('root', 'root'), use_http=True)