+++ tests/integration_tests/test_ws_connection.py
@@ -273,7 +273,7 @@ class TestWebSocketConnection(TestCase):
             opts = {"id": uid, "author": uid, "title": uid, "text": uid}
             connection.create("ws_article", opts)
             connection.create("ws_article2", opts)
-            time.sleep(0.1)
+            time.sleep(0.3)
             self.assertEqual(a_list[0]['result']['action'], 'CREATE')
             self.assertEqual(a_list[1]['result']['action'], 'CREATE')
             self.assertEqual(a_list[0]['result']['result'], {**opts, "id": f"ws_article:{uid}"})
@@ -284,8 +284,7 @@ class TestWebSocketConnection(TestCase):
         with surreal.connect() as connection:
             connection.use("test", "test")
             res = connection.count("not_exists")
-            self.assertFalse(res.is_error())
-            self.assertEqual(0, res.result)
+            self.assertTrue(res.is_error())
             self.assertEqual("SELECT count() FROM not_exists GROUP ALL;", res.query)
             count = connection.count("article").result
             uid = get_random_series(6)
@@ -293,13 +292,12 @@ class TestWebSocketConnection(TestCase):
             new_count = connection.count("article").result
             self.assertEqual(new_count, count + 1)
 
-    def test_count_is_zero_if_wrong(self):
+    def test_count_is_error_if_wrong(self):
         surreal = Surreal(URL, credentials=('root', 'root'))
         with surreal.connect() as connection:
             connection.use("test", "test")
             res = connection.count("wrong")
-            self.assertFalse(res.is_error())
-            self.assertEqual(0, res.result)
+            self.assertTrue(res.is_error())
 
     def test_count_returns_fields(self):
         surreal = Surreal(URL, credentials=('root', 'root'))